requirements.txt
fastapi
uvicorn
python-multipart
chromadb
langchain
langchain-google-genai
langchain-community
pypdf
sentence-transformers
python-dotenv